feat: fallback for /gif random
gif random logic is basically, fetch a random word and pass it as query
param to Giphy API.

It had a tendency to fail on fancy words such as 'preprosperous'

Now, if the Giphy query don't return anything, we'll do another query
for random cat gif.

Cause everybody loves cats, and they come in large supplies.